+++ tests/cli_rgr.rs
@@ -63,7 +63,10 @@ fn command_md() -> anyhow::Result<()> {
     let stdout = String::from_utf8(output.stdout).unwrap();
 
     assert_eq!(stdout.lines().count(), 5);
-    assert!(stdout.contains("| -----: | :--------: | --------------- |"), "separator");
+    assert!(
+        stdout.contains("| -----: | :--------: | --------------- |"),
+        "separator"
+    );
     assert!(stdout.contains("| 130218 |  ctg:I:2   | I:100001-230218 |"));
 
     let mut cmd = Command::cargo_bin("rgr")?;
@@ -78,7 +81,10 @@ fn command_md() -> anyhow::Result<()> {
     let stdout = String::from_utf8(output.stdout).unwrap();
 
     assert_eq!(stdout.lines().count(), 5);
-    assert!(stdout.contains("| ---------: | ---------- | --------------- |"), "separator");
+    assert!(
+        stdout.contains("| ---------: | ---------- | --------------- |"),
+        "separator"
+    );
     assert!(stdout.contains("| 130,218.00 | ctg:I:2    | I:100001-230218 |"));
 
     Ok(())